pyproject.toml
[project]
requires-python = ">=3.9"
name = "sentiment-gemini-api"
version = "0.1.0"
description = "A short description of the project"
authors = [
    { name = "Your name", email = "you@example.com" },
]
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn==0.23.2",
    "pydantic>=2.0.0",
    "requests>=2.32.0",
    "loguru>=0.7.0",
    "joblib>=1.2.0",
    "scikit-learn>=1.5.0",
    "google-generativeai>=0.5.4",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2",
    "black>=24.3",
    "autopep8>=2.0.0",
    "ipdb>=0.13.0",
    "pylint>=3.0.0"
]
aws = [
    "mangum>=0.17.0"
]

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]